COTECH-787 refactor the bids refreshing change
